+++ package.json
@@ -23,6 +23,7 @@
     "@mui/icons-material": "^5.14.16",
     "@mui/material": "^5.14.16",
     "@reduxjs/toolkit": "^1.9.7",
+    "axios": "^1.6.2",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
