+++ src/components/AddDishForm/AddDishFormImageUpload/index.js
@@ -1,0 +1,1 @@
+export { AddDishFormImageUpload } from './AddDishFormImageUpload';
