+++ src/components/AddDishForm/AddDishFormNavButtons/index.js
@@ -1,0 +1,1 @@
+export { AddDishFormNavButtons } from './AddDishFormNavButtons';
