+++ src/components/AddDishForm/AddDishFormSpiceLevel/index.js
@@ -1,0 +1,1 @@
+export { AddDishFormSpiceLevel } from './AddDishFormSpiceLevel';
