+++ src/components/AddDishForm/AddDishFromStepper/index.js
@@ -1,0 +1,1 @@
+export { AddDishFormStepper } from './AddDishFromStepper';
