+++ src/components/AddDishForm/AddDishFromSteps/AddDishFormStepFour/index.js
@@ -1,0 +1,1 @@
+export { AddDishFormStepFour } from './AddDishFormStepFour';
