+++ src/components/AddDishForm/AddDishFromSteps/AddDishFormStepOne/index.js
@@ -1,0 +1,1 @@
+export { AddDishFromStepOne } from './AddDishFormStepOne';
