+++ src/components/AddDishForm/AddDishFromSteps/AddDishFormStepThree/index.js
@@ -1,0 +1,1 @@
+export { AddDishFormStepThree } from './AddDishFormStepThree';
