+++ src/components/AddDishForm/AddDishFromSteps/AddDishFormStepTwo/index.js
@@ -1,0 +1,1 @@
+export { AddDishFormStepTwo } from './AddDishFormStepTwo';
