+++ src/components/AddDishForm/AddDishFromSteps/index.js
@@ -1,0 +1,4 @@
+export { AddDishFormStepFour } from './AddDishFormStepFour';
+export { AddDishFromStepOne } from './AddDishFormStepOne';
+export { AddDishFormStepThree } from './AddDishFormStepThree';
+export { AddDishFormStepTwo } from './AddDishFormStepTwo';
