+++ src/components/AddDishForm/index.js
@@ -1,0 +1,3 @@
+export { AddDishFormImageUpload } from './AddDishFormImageUpload';
+export { AddDishFormNavButtons } from './AddDishFormNavButtons';
+export { AddDishFormSpiceLevel } from './AddDishFormSpiceLevel';
